+++ NEWS.md
@@ -6,6 +6,13 @@ For more fine-grained list of changes or to report a bug, consult
 * [The issues log](https://github.com/ropensci/RNeXML/issues)
 * [The commit log](https://github.com/ropensci/RNeXML/commits/master)
 
+
+v2.4.5
+------
+
+- CRAN-requested patch for collision with {cli} namespace (#251)
+- taxadb -> taxalight for fast resolution of taxon ids
+
 v2.4.4
 ------